requirements.txt
Flask
pyserial
setuptools
